Move missing_mods_panel.py into RimSort.windows
