+++ RimSort/views/main_content_panel.py
@@ -74,7 +74,7 @@ from util.steam.steamworks.wrapper import (
 from util.todds.wrapper import ToddsInterface
 from util.xml import json_to_xml_write, xml_path_to_json
 from RimSort.views.game_configuration_panel import GameConfiguration
-from window.missing_mods_panel import MissingModsPrompt
+from RimSort.windows.missing_mods_panel import MissingModsPrompt
 from window.rule_editor_panel import RuleEditor
 from RimSort.windows.runner_panel import RunnerPanel
 from window.workshop_mod_updater_panel import ModUpdaterPrompt
